fixed tests and updated notebook
