+++ composer/datasets/streaming/download.py
@@ -12,45 +12,51 @@ import urllib.parse
 from typing import Optional
 
 from composer.datasets.streaming.format import split_compression_suffix
-# TODO: refactor to use object store for download, until then, use this private method.
+from composer.utils.object_store.object_store import ObjectStore
 from composer.utils.object_store.s3_object_store import S3ObjectStore
 from composer.utils.object_store.sftp_object_store import SFTPObjectStore
 
-__all__ = ['download_or_wait']
+__all__ = ['download_or_wait', 'get_object_store']
 
 
-def download_from_s3(remote: str, local: str) -> None:
-    """Download a file from remote to local.
+def get_object_store(remote: str) -> ObjectStore:
+    """Use the correct download handler to download the file
 
     Args:
-        remote (str): Remote path (S3).
-        local (str): Local path (local filesystem).
-        timeout (float): How long to wait for shard to download before raising an exception.
+        remote (Optional[str]): Remote path (local filesystem).
+        timeout (float): How long to wait for file to download before raising an exception.
     """
+    if remote.startswith('s3://'):
+        return _get_s3_object_store(remote)
+    elif remote.startswith('sftp://'):
+        return _get_sftp_object_store(remote)
+    else:
+        raise ValueError('unsupported upload scheme')
+
+
+def _get_s3_object_store(remote: str) -> S3ObjectStore:
     obj = urllib.parse.urlparse(remote)
     if obj.scheme != 's3':
         raise ValueError(f"Expected obj.scheme to be 's3', got {obj.scheme} for remote={remote}")
-    bucket, key = obj.netloc, obj.path.lstrip('/')
+    bucket = obj.netloc
     object_store = S3ObjectStore(bucket=bucket)
-    object_store.download_object(key, local)
+    return object_store
 
 
-def download_from_sftp(remote: str, local: str) -> None:
-    """Download a file from remote SFTP server to local filepath.
+def _get_sftp_object_store(remote: str) -> SFTPObjectStore:
+    # Get SSH key file if specified
+    key_filename = os.environ.get('COMPOSER_SFTP_KEY_FILE', None)
+    known_hosts_filename = os.environ.get('COMPOSER_SFTP_KNOWN_HOSTS_FILE', None)
 
-    Authentication must be provided via username/password in the ``remote`` URI, or a valid SSH config, or a default key
-    discoverable in ``~/.ssh/``.
+    object_store = SFTPObjectStore(
+        host=remote,
+        known_hosts_filename=known_hosts_filename,
+        key_filename=key_filename,
+    )
+    return object_store
 
-    Args:
-        remote (str): Remote path (SFTP).
-        local (str): Local path (local filesystem).
-    """
-    url = urllib.parse.urlsplit(remote)
-    remote_path = url.path
 
-    object_store = SFTPObjectStore(host=remote)
-
-    object_store.download_object(remote_path, local)
+__all__ = ['download_or_wait']
 
 
 def download_from_local(remote: str, local: str) -> None:
@@ -85,10 +91,11 @@ def dispatch_download(remote: Optional[str], local: str):
 
     if not remote:
         raise ValueError('In the absence of local dataset, path to remote dataset must be provided')
-    elif remote.startswith('s3://'):
-        download_from_s3(remote, local)
-    elif remote.startswith('sftp://'):
-        download_from_sftp(remote, local)
+    elif remote.startswith('s3://') or remote.startswith('sftp://'):
+        url = urllib.parse.urlsplit(remote)
+        remote_path = url.path
+        object_store = get_object_store(remote)
+        object_store.download_object(remote_path, local)
     else:
         download_from_local(remote, local)
