+++ tests/datasets/test_streaming.py
@@ -47,7 +47,7 @@ def get_fake_samples_decoders(num_samples: int) -> Tuple[List[Dict[str, bytes]],
 def write_synthetic_streaming_dataset(dirname: str,
                                       samples: List[Dict[str, bytes]],
                                       shard_size_limit: int,
-                                      compression: Optional[str] = StreamingDatasetWriter.default_compression,
+                                      compression: Optional[str] = None,
                                       upload: Optional[str] = None) -> None:
     first_sample_fields = list(samples[0].keys())
     with StreamingDatasetWriter(dirname=dirname,
